Changed release script to detect Petit Larceny more reliably.
git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5652 c9681fa6-b8ae-40c0-a6da-707a35ad5973

+++ test/Scripts/package-bin-release.sh
@@ -53,12 +53,13 @@ if [ -e larceny.np ]; then
 fi
 
 # Remove the include directory from the non-Petit distributions.
-if [ ! -e petit.bin ] && [ ! -e petit.bin.exe ]; then
+if [ ! -e petit.bin ] && [ ! -e petit-larceny.bin ] && [ ! -e petit.bin.exe ];
+  then
     rm -rf include
 fi
 
 # For Petit:
-if [ -e petit.bin ] || [ -e petit.bin.exe ]
+if [ -e petit.bin ] || [ -e petit-larceny.bin ] || [ -e petit.bin.exe ]
 #    remove the .o files
 then find . -name '*.o' -type f | xargs rm -f
 #    remove the generated .c files
